refact: Money에서 불필요한 박스 타입 및 Float에서 Double로 변경

+++ src/main/java/domains/Money.java
@@ -36,13 +36,13 @@ public class Money {
         return amount / LOTTO_PRICE;
     }
 
-    public Float calculateRate(List<Rank> rankList) {
-        if (amount == 0) return (float) 0;
+    public double calculateRate(List<Rank> rankList) {
+        if (amount == 0) return  0.0d;
 
         long totalWinningMoney = rankList.stream()
                 .mapToLong(Rank::getWinningMoney)
                 .sum();
 
-        return (float) totalWinningMoney / amount;
+        return (double) totalWinningMoney / amount;
     }
 }
